pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "field-assist-bot"
version = "0.1.0"
description = "AI-powered Discord bot for field research operations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.5.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.0",
  "python-dotenv>=1.0.1",
  "structlog>=24.4.0",
  "sqlalchemy[asyncio]>=2.0.36",
  "alembic>=1.14.0",
  "aiosqlite>=0.20.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.32.0",
  "httpx>=0.28.0",
  "apscheduler>=3.10.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.8.0",
  "mypy>=1.13.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "D"]
ignore = ["D100", "D104"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

